refactor: moving unnecessary logic from PostList component to App.js
